Create integration router in builder
