+++ .changeset/small-hotels-shake.md
@@ -1,5 +1,5 @@
 ---
-'@backstage/plugin-catalog-backend': minor
+'@backstage/plugin-catalog-backend': patch
 ---
 
 Add authorization to catalog-backend entities GET endpoints
