+++ plugins/catalog-backend/src/service/NextRouter.ts
@@ -17,23 +17,16 @@
 import { errorHandler } from '@backstage/backend-common';
 import {
   analyzeLocationSchema,
-  Entity,
   locationSpecSchema,
-  parseEntityRef,
   stringifyEntityRef,
 } from '@backstage/catalog-model';
 import { Config } from '@backstage/config';
 import { NotFoundError } from '@backstage/errors';
-import { RESOURCE_TYPE_CATALOG_ENTITY } from '@backstage/plugin-catalog-common';
-import {
-  createPermissionIntegrationRouter,
-  PermissionRule,
-} from '@backstage/plugin-permission-node';
 import express from 'express';
 import Router from 'express-promise-router';
 import { Logger } from 'winston';
 import yn from 'yn';
-import { EntitiesCatalog, EntitiesSearchFilter } from '../catalog';
+import { EntitiesCatalog } from '../catalog';
 import { LocationAnalyzer } from '../ingestion/types';
 import {
   basicEntityFilter,
@@ -51,7 +44,7 @@ export interface NextRouterOptions {
   refreshService?: RefreshService;
   logger: Logger;
   config: Config;
-  permissionRules?: PermissionRule<Entity, EntitiesSearchFilter, unknown[]>[];
+  permissionIntegrationRouter?: express.Router;
 }
 
 export async function createNextRouter(
@@ -64,7 +57,7 @@ export async function createNextRouter(
     refreshService,
     config,
     logger,
-    permissionRules,
+    permissionIntegrationRouter,
   } = options;
 
   const router = Router();
@@ -88,16 +81,12 @@ export async function createNextRouter(
     });
   }
 
+  if (permissionIntegrationRouter) {
+    router.use(permissionIntegrationRouter);
+  }
+
   if (entitiesCatalog) {
     router
-      .use(
-        createPermissionIntegrationRouter({
-          resourceType: RESOURCE_TYPE_CATALOG_ENTITY,
-          getResource: resourceRef =>
-            getEntityResource(resourceRef, entitiesCatalog),
-          rules: permissionRules ?? [],
-        }),
-      )
       .get('/entities', async (req, res) => {
         const { entities, pageInfo } = await entitiesCatalog.entities({
           filter: parseEntityFilterParams(req.query),
@@ -205,23 +194,6 @@ export async function createNextRouter(
 
   router.use(errorHandler());
   return router;
-}
-
-async function getEntityResource(
-  resourceRef: string,
-  entitiesCatalog: EntitiesCatalog,
-): Promise<Entity | undefined> {
-  const parsed = parseEntityRef(resourceRef);
-
-  const { entities } = await entitiesCatalog.entities({
-    filter: basicEntityFilter({
-      kind: parsed.kind,
-      'metadata.namespace': parsed.namespace,
-      'metadata.name': parsed.name,
-    }),
-  });
-
-  return entities[0];
 }
 
 function getBearerToken(
